sdm710-common: decommonize sensors blobs
Sensors HAL are tightly connected with camera HAL. Due to facts that
sirius probably will not receive A11 update and camera HAL is highly
device specific we cannot update sensors HAL in sdm710-common.
We have to decommonize sensors HAL to be able to update camera HAL
on other devices.